Исправление ошибок

+++ controllers/cards.js
@@ -1,12 +1,11 @@
 const Card = require("../models/card.js");
-
-const DEFAULT_ERROR_CODE = 500;
-const NOT_CORRECT_ERROR_CODE = 400;
-const NOT_FOUND_ERROR_CODE = 404;
-
-const handleError = (res, statusCode, message) => {
-  res.status(statusCode).send({ message });
-};
+const {
+  DEFAULT_ERROR_CODE,
+  NOT_CORRECT_ERROR_CODE,
+  NOT_FOUND_ERROR_CODE,
+  handleError,
+  isValidId,
+} = require("../utils/utils.js");
 
 module.exports.getCards = (req, res) => {
   Card.find({})
@@ -18,7 +17,13 @@ module.exports.createCard = (req, res) => {
   const { name, link } = req.body;
   const userId = req.user._id;
 
-  if (!name || name.length < 2 || name.length > 30 || !link) {
+  if (
+    !name ||
+    name.length < 2 ||
+    name.length > 30 ||
+    !link ||
+    !isValidId(userId)
+  ) {
     return handleError(
       res,
       NOT_CORRECT_ERROR_CODE,
@@ -34,13 +39,12 @@ module.exports.createCard = (req, res) => {
 module.exports.getCardById = (req, res) => {
   const cardId = req.params.cardId;
 
+  if (!isValidId(cardId)) {
+    return handleError(res, NOT_FOUND_ERROR_CODE, "Карточка не найдена");
+  }
+
   Card.findById(cardId)
-    .then((card) => {
-      if (!card) {
-        return handleError(res, NOT_FOUND_ERROR_CODE, "Карточка не найдена");
-      }
-      res.send({ data: card });
-    })
+    .then((card) => res.send({ data: card }))
     .catch((err) => handleError(res, DEFAULT_ERROR_CODE, err.message));
 };
 
@@ -48,17 +52,16 @@ module.exports.addLike = (req, res) => {
   const { cardId } = req.params;
   const userId = req.user._id;
 
+  if (!isValidId(cardId)) {
+    return handleError(res, NOT_FOUND_ERROR_CODE, "Карточка не найдена");
+  }
+
   Card.findByIdAndUpdate(
     cardId,
     { $addToSet: { likes: userId } },
     { new: true }
   )
-    .then((card) => {
-      if (!card) {
-        return handleError(res, NOT_FOUND_ERROR_CODE, "Карточка не найдена");
-      }
-      res.send({ data: card });
-    })
+    .then((card) => res.send({ data: card }))
     .catch((err) => handleError(res, DEFAULT_ERROR_CODE, err.message));
 };
 
@@ -66,25 +69,23 @@ module.exports.removeLike = (req, res) => {
   const { cardId } = req.params;
   const userId = req.user._id;
 
+  if (!isValidId(cardId)) {
+    return handleError(res, NOT_FOUND_ERROR_CODE, "Карточка не найдена");
+  }
+
   Card.findByIdAndUpdate(cardId, { $pull: { likes: userId } }, { new: true })
-    .then((card) => {
-      if (!card) {
-        return handleError(res, NOT_FOUND_ERROR_CODE, "Карточка не найдена");
-      }
-      res.send({ data: card });
-    })
+    .then((card) => res.send({ data: card }))
     .catch((err) => handleError(res, DEFAULT_ERROR_CODE, err.message));
 };
 
 module.exports.deleteCard = (req, res) => {
   const cardId = req.params.cardId;
 
+  if (!isValidId(cardId)) {
+    return handleError(res, NOT_FOUND_ERROR_CODE, "Карточка не найдена");
+  }
+
   Card.findByIdAndRemove(cardId)
-    .then((card) => {
-      if (!card) {
-        return handleError(res, NOT_FOUND_ERROR_CODE, "Карточка не найдена");
-      }
-      res.send({ data: card });
-    })
+    .then((card) => res.send({ data: card }))
     .catch((err) => handleError(res, DEFAULT_ERROR_CODE, err.message));
 };
